fix: cast BF16 to float32 before numpy, disable torch.compile

+++ src/trainer.py
@@ -348,7 +348,7 @@ class Trainer:
             total += labels.size(0)
             correct += predicted.eq(labels).sum().item()
 
-            all_probs.append(probs.cpu().numpy())
+            all_probs.append(probs.float().cpu().numpy())
             all_labels.append(labels.cpu().numpy())
 
         avg_loss = total_loss / len(loader)
